refine daily stock api to get all data
add a analyze tool to download stock daily prices

+++ cmd/notifier/main.go
@@ -40,6 +40,6 @@ func main() {
 		if err != nil {
 			log.Fatalf("upload image with error=%v", err)
 		}
-		bot.Send(&tg.User{ID: user.UserID}, link)
+		bot.Send(&tg.User{ID: int64(user.UserID)}, link)
 	}
 }
